Исправляет работу событий на открытие и отображение большого изображения
