+++ js/mocks/mock-values.json
@@ -1,0 +1,21 @@
+{
+	"messages": [
+		"Всё отлично!",
+		"В целом всё неплохо. Но не всё.",
+		"Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.",
+		"Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.",
+		"Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.",
+		"Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!"
+	],
+	"names": [
+		"Серега Серый",
+		"Саня Белый",
+		"Никитос Черный",
+		"Гошан Фиолетовый",
+		"Леха Бордовый",
+		"Максон Зеленый",
+		"Дэнчик Желтый",
+		"Ромчик Пурпурный",
+		"Викуся Камуфляжная"
+	]
+}
