+++ js/render-data.ts
@@ -1,0 +1,6 @@
+import { createMocks } from './mocks/mock.ts';
+import { renderPictures } from './render-pictures.ts';
+
+const usersData = createMocks();
+
+renderPictures(usersData);
