+++ js/utils.ts
@@ -7,16 +7,33 @@ const getRandomInteger = (a: number, b: number) : number => {
 
 const getRandomArrayElement = <El>(elements: El[] | readonly El[]): El => elements[getRandomInteger(0, elements.length - 1)];
 
-const isPicturePressedByClick = (evt: MouseEvent) => {
-	const eventElement = <HTMLElement>evt.target;
-	return eventElement!.classList.contains('picture__img');
-};
-
-const isPicturePressedByKeydown = (evt: KeyboardEvent) => {
-	const eventElement = <HTMLElement>evt.target;
-	return eventElement!.classList.contains('picture') && evt!.key === 'Enter';
-};
-
 const isEscapeKey = (evt: KeyboardEvent) => evt.key === 'Escape';
 
-export {getRandomInteger, getRandomArrayElement, isPicturePressedByClick, isPicturePressedByKeydown, isEscapeKey};
+const findTemplate = <E extends HTMLElement = HTMLElement>(id: string) => {
+	const templateElement = document.querySelector<HTMLTemplateElement>(`#${id}`);
+	if (templateElement === null) {
+		throw new Error(`Template with id ${id} not found`);
+	}
+
+	const template = templateElement.content.firstElementChild;
+	if (template === null) {
+		throw new Error(`Template with id ${id} not found`);
+	}
+
+	return template as E;
+};
+
+const findBEMElement = <E extends HTMLElement = HTMLElement>(blockNode: Element, element: string, block?: string) => {
+	if (!block) {
+		block = blockNode.classList[0];
+	}
+
+	const elementNode = blockNode.querySelector<E>(`.${block}__${element}`);
+	if (elementNode === null) {
+		throw new Error(`Element ${element} not found in block ${block}`);
+	}
+
+	return elementNode;
+};
+
+export {getRandomInteger, getRandomArrayElement, isEscapeKey, findTemplate, findBEMElement};
